requirements.txt
asgiref==3.4.1
attrs==21.2.0
boto3==1.20.31
certifi==2021.5.30
charset-normalizer==2.0.4
cloudinary==1.26.0
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.5
Django==3.2.5
django-cors-headers==3.7.0
django-environ==0.6.0
django-filter==2.4.0
django-redis==5.0.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
drf-spectacular==0.17.2
Faker==8.10.0
gunicorn==20.1.0
idna==3.2
inflection==0.5.1
jsonschema==3.2.0
Pillow==8.3.1
psycopg2-binary==2.9.1
PyJWT==2.1.0
pyrsistent==0.18.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.26.0
sentry-sdk==1.3.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.26.6
whitenoise==5.3.0
